重构组件，支持设置样式
