--- src/components/index.js
@@ -1,12 +1,0 @@
-// Web Components
-
-import "./section.js";
-import "./panel.js";
-import "./list.js";
-import "./item.js";
-import "./select.js";
-import "./option.js";
-import "./switch.js";
-import "./button.js";
-import "./text.js";
-import "./divider.js";
